個人別粗利集計表の反映
