+++ src/eventHandlers/formGrossProfitTable/results/Results.tsx
@@ -7,10 +7,10 @@ import {useAndpadProcurement} from '../../hooks/useAndpadProcurement';
 import {useProjTypes} from '../../hooks/useProjTypes';
 import {getDatePeriod} from './helper/getDatePeriod';
 import {type SummaryContracts, getSummaryContracts} from '../../helpers/getSummaryContracts';
-import {Stack} from '@mui/material';
+import {LinearProgress, Stack} from '@mui/material';
 import {CumulativeTable} from './cumulativeTable/CumulativeTable';
-import {useMembers} from '@/eventHandlers/hooks/useMembers';
 import {type IEmployees} from '@api/getEmployees';
+import {useMembers} from '@/eventHandlers/hooks/useMembers';
 
 export const Results = () => {
 	const [
@@ -40,13 +40,13 @@ export const Results = () => {
 	const {data: projTypes} = useProjTypes();
 	const {data: members} = useMembers({area});
 
+	const isLoading = !projects
+		|| !contracts
+		|| !andpadProcurement
+		|| !projTypes;
+
 	const summaryContracts = useMemo(() => {
-		if (
-			!projects
-			|| !contracts
-			|| !andpadProcurement
-			|| !projTypes
-		) {
+		if (isLoading) {
 			return [] as SummaryContracts[];
 		}
 
@@ -56,24 +56,33 @@ export const Results = () => {
 			andpadProcurement,
 			projTypes,
 		});
-	}, [projects, contracts, andpadProcurement, projTypes]);
+	}, [projects, contracts, andpadProcurement, projTypes, isLoading]);
 
 	return (
-		<Stack spacing={2}>
-			<CumulativeTable
-				area={area}
-				periods={periods}
-				year={year}
-				summaryContracts={summaryContracts}
-				employeesNum={members ? members.length : 0}
-			/>
-			<GrossProfitByPerson
-				area={area}
-				members={members || [] as IEmployees[]}
-				periods={periods}
-				year={year}
-				summaryContracts={summaryContracts}
-			/>
+		<Stack
+			spacing={2}
+			px={4}
+			pt={2}
+			pb={4}
+		>
+			{!isLoading
+				&& <>
+					<CumulativeTable
+						area={area}
+						periods={periods}
+						year={year}
+						summaryContracts={summaryContracts}
+						employeesNum={members ? members.length : 0}
+					/>
+					<GrossProfitByPerson
+						area={area}
+						members={members || [] as IEmployees[]}
+						periods={periods}
+						year={year}
+						summaryContracts={summaryContracts}
+					/>
+				</>}
+			{isLoading && <LinearProgress color='primary' />}
 		</Stack>
 	);
 };
