+++ src/eventHandlers/formGrossProfitTable/results/helper/contractsFormatGrossProfitTable.ts
@@ -10,9 +10,11 @@ import Big from 'big.js';
 export const contractsFormatGrossProfitTable = ({
 	filteredContracts,
 	monthsNum,
+	cocoConst = '',
 }: {
 	filteredContracts: SummaryContracts[];
 	monthsNum: number;
+	cocoConst?: string;
 }) => {
 	const formattingContracts = filteredContracts.reduce<Record<ProjTypeList, GrossProfitTableRow>>((acc, {
 		projTypeForTotalization,
@@ -29,6 +31,7 @@ export const contractsFormatGrossProfitTable = ({
 				grossprofitAmtTotal: Number(grossProfitAmount),
 				grossProfitCoco: Number(grossProfitAmtCoco),
 				introFeeYume: Number(introFeeYume),
+				cocoConst,
 			};
 		} else {
 			acc[projTypeForTotalization] = {
@@ -37,6 +40,7 @@ export const contractsFormatGrossProfitTable = ({
 				grossprofitAmtTotal: acc[projTypeForTotalization].grossprofitAmtTotal + Number(grossProfitAmount),
 				grossProfitCoco: acc[projTypeForTotalization].grossProfitCoco + Number(grossProfitAmtCoco),
 				introFeeYume: acc[projTypeForTotalization].introFeeYume + Number(introFeeYume),
+				cocoConst,
 			};
 		}
 
@@ -48,6 +52,7 @@ export const contractsFormatGrossProfitTable = ({
 		if (formattingContracts[projType] === undefined) {
 			formattingContracts[projType] = {
 				...grossProfitTblRowInit,
+				cocoConst,
 				projType,
 			};
 		} else {
